% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyCellEmpty.R
\name{anyCellEmpty}
\alias{anyCellEmpty}
\title{Check if any cell of a vector is empty}
\usage{
anyCellEmpty(x, flag = c(0, NA, -999))
}
\arguments{
\item{x}{table (data.frame/2D-matrix) to be checked}

\item{flag}{What is meant to be "empty". \code{flag} could be a vector but all of it elements must be different of the posible valid values of \code{x}.}
}
\details{
See \link{stats::complete.cases} for data.frames/matrix. incompleteBlocks <- (1:nrow(x))[!complete.cases(x)]
}
\examples{
xe <- c(3,3,3, 0, -999, NA)
anyCellEmpty(xe, flag = c(0, NA, -999))
xe <- 1:5
anyCellEmpty(xe)
}
\seealso{
if in the example
x[ complete.cases(x),]
x[!complete.cases(x),]
and the function EmptyRow.R
}
