% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkClass.R
\name{checkClass}
\alias{checkClass}
\title{look for non-numeric columns and correct them to numeric}
\usage{
checkClass(x, cols = NULL)
}
\arguments{
\item{x}{A data.frame (not matrix).}

\item{cols}{specific columns of \code{x}. If \code{NULL} (default), the complete \code{x} is class-checked.}
}
\value{
a numeric vector, data.frame or matrix with
}
\details{
look for non-numeric columns and correct them to numeric by the function \link{rmNonNumeric}
}
\examples{
checkClass(head(iris))
print(head(iris))
}
