% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckRowSums.R
\name{CheckRowSums}
\alias{CheckRowSums}
\title{Checks that evey row sums \code{total}}
\usage{
CheckRowSums(x, total = 100, cols = NULL)
}
\arguments{
\item{x}{data.frame or matrix}

\item{total}{The true total value.}

\item{cols}{integer vector with the columns of \code{x} to be checked.}
}
\value{
a logical vector with \code{TRUE} in the rows wich are equal to total
}
\examples{
CheckRowSums(iris[1:5,1:3], total = 10)
CheckRowSums(iris[1:5,1:3], total = 10, cols = c(1,2))
CheckRowSums(iris[1:5,1:3], total = rowSums(iris[1:5,1:3]))
}
