% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countEmptyCells.R
\name{countEmptyCells}
\alias{countEmptyCells}
\title{count the number of empty cells in each row in order to check for large missing values in a row.}
\usage{
countEmptyCells(x, ...)
}
\arguments{
\item{x}{data.frame or matrix}

\item{...}{further argumente passed to \link{setEmptyCellsValue}}
}
\details{
count the number of empty cells in each row in order to check for large missing values in a row.
}
\examples{
xe <- matrix(data = NA, nrow = 4, ncol = 3)
xe[1, ] <- 1
xe[2, 2:3] <- 2
xe[3,   3] <- 3
print(xe)
countEmptyCells(xe)

xe <- matrix(data = -99, nrow = 4, ncol = 3)
xe[1, ] <- 1
xe[2, 2:3] <- 2
xe[3,   3] <- 3
print(xe)
countEmptyCells(xe, ef = -99)
}
