% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmptyRow.R
\name{EmptyRow}
\alias{EmptyRow}
\title{Check if any row of a table is empty}
\usage{
EmptyRow(x, ef = c(NA, -999, 0))
}
\arguments{
\item{x}{table (data.frame/2D-matrix) to be checked for completely empty rows.}

\item{...}{\code{ef} arguments passed to \link{setEmptyCellsValue}.}
}
\value{
a list with \code{TF}, a logical vector with TRUE value corresponding with complete empty rows, and \code{ID} giving the row number of the empty rows.
}
\details{
Check if any row of a table is empty
}
\examples{
flag <- NA
xe <- data.frame(a = 3:5, b = flag, c = flag)
xe[2, 1] <- flag
xe[1, 2] <- 0
xe[1, 1] <- 0
xe[3, 2] <- 3
print(xe)
EmptyRow(x = xe, ef = c(flag, 0))
}
\seealso{
if in the example \code{flag = NA}

x[ complete.cases(x),]

x[!complete.cases(x),]
}
