% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPosIndex.R
\name{getPosIndex}
\alias{getPosIndex}
\title{Look for the index in an R object}
\usage{
getPosIndex(cnames, sc, ec = NULL)
}
\arguments{
\item{cnames}{\code{c}olumn \code{names} of the entire table}

\item{sc}{names of the \code{s}tarting \code{c}olumn of the sub-table}

\item{ec}{names of the \code{e}nd \code{c}olumn column of the sub-table}
}
\value{
An integer for the position (column number in case of data.frame or matrix) of a subset indicated by (column, row, etc.) names
}
\examples{
names(iris)
getPosIndex(cnames = names(iris),
                      sc = "Sepal.Width", ec = "Petal.Width")
getPosIndex(cnames = names(iris),
                      sc = "Sepal.Width")
}
