% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcq.R
\name{mcq}
\alias{mcq}
\title{Multiple choice questions (mcq) check.}
\usage{
mcq(x, opts, inv_fw = NA, fw = NA)
}
\arguments{
\item{x}{table (data.frame/matrix) from data comes.}

\item{opts}{options}

\item{inv_fw}{change the invalid options with \code{inv_fw}. Default to \code{NA}. If \code{NULL} it keeps the original values.}

\item{fw}{fillwith. If \code{inv_fw = NA} it will be overwriten by \code{fw}.}

\item{fc}{fill with to check visually}
}
\value{
\code{x} with elements different of \code{opts} or blanks (NA) filled with \code{fc}, and \code{NA}'s are filled with \code{fw}
}
\details{
All \code{x}, \code{opts}, \code{fw} and \code{fc} must be the same class
}
\examples{
xe <- c(3, "o.9", 8)
mcq(xe, 1:3)

xe <- c(3, 0, 8, NA)
mcq(x = xe, opts = 0:3, fw = -99.009)
mcq(x = xe, opts = 0:3, fw = NA, inv_fw = NULL)
mcq(x = xe, opts = 0:3)
}
