% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoveCells2left.R
\name{MoveCells2left}
\alias{MoveCells2left}
\title{Move filled cells to the leftmost empty cells.}
\usage{
MoveCells2left(df, fillwith = NA, shift = 1, ef = c(NA, -999, 0))
}
\arguments{
\item{df}{table (data.frame/2D-matrix) to be checked}

\item{fillwith}{what is the character/number to fill the "empty" cells. Default the same as \code{flag}}

\item{...}{\code{ef} arguments passed to \link{setEmptyCellsValue}}
}
\value{
a data.frame/2D-matrix
}
\examples{
# example 1
xe <- data.frame(x=c("l","m",NA,NA,"p"),
                 y=c(NA,"b","c",NA,NA),
                 z=c("u",NA,"w","x","y"))
print(xe)
MoveCells2left(df = xe, ef = NA)
# example 2
xe <- data.frame(a = 3:6, b = 0, c = 0, d = 0, e = 0)
xe[2, 1] <- 0
xe[1, 4] <- 3
xe[2, 3] <- 3
xe[3, 3] <- 3
xe[3, 5] <- 3
xe[4, 2] <- 3
print(xe)
MoveCells2left(df = xe, ef = 0)
}
\references{
http://stackoverflow.com/questions/23285215/advanced-dataframe-manipulation-in-r-shifting-cells-to-the-left
}
