% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSums2.R
\name{mSums2}
\alias{mSums2}
\title{Check and correct Multiple sums}
\usage{
mSums2(x, sc, ec, scT, ecT, nr = 1, nc = 1, ...)
}
\arguments{
\item{x}{Table to be checked. data not need to be contigous but must be equally spaced.}

\item{sc, scT}{names of the \code{s}tarting \code{c}olumn of the table \code{x} for data to be checked \code{sc} and for the totals, \code{scT}}

\item{ec, scT}{names of the \code{e}nd \code{c}olumn column of the table \code{x} for data to be checked \code{sc} and for the totals, \code{scT}}

\item{nr}{Block length}

\item{nc}{number of blocks. Default to 1, i.e., a the totals must be a vector.}

\item{...}{further arguments passed to \link{ccSums}}
}
\value{
A list with
\describe{
  \item{x}{input \code{x} but data checked}
  \item{cPos}{column numbers representing \code{sc} and \code{ec} of data \code{x}}
  \item{cPosT}{column numbers representing \code{sc} and \code{ec} of the totals}
}
}
\details{
\code{total} columns must be contiguous, and data blocks too.
}
\examples{
xe <- as.data.frame(matrix(c(-1:-5, 33301:33305,
                             -1:-5, 33301:33305,
                             1:10), # totals
                           nrow = 5, ncol = 6))
flag <- NA
xe[1, 1] <- flag
xe[2, 1] <- flag
xe[2, 3] <- flag
print(xe)
mSums2(xe, sc = "V1", ec = "V4",
      scT = "V5", ecT = "V6",
      nr = 2, nc = 2, ef = -99.009)
}
