% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q12_2.R
\name{q12_2}
\alias{q12_2}
\title{Verify question 12.2 of comercio.xlsx}
\usage{
q12_2(x, skipFlag, opts)
}
\arguments{
\item{x}{data.frame or matrix}

\item{skipFlag}{what to skip}

\item{opts}{valid options}
}
\examples{
cpos <- getPosIndex(cnames = names(df1),
                    sc = "P12_2_1", ec = "P12_2_0")
q12_2(x = df1[, cpos], skipFlag = 6, opts = 1:6)
}
