% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q15_1.R
\name{q15_1}
\alias{q15_1}
\title{Does the job in question 15 in comercio.xlsx}
\usage{
q15_1(x, sc, ec, nr = 1, nc = 1, cPos, ...)
}
\arguments{
\item{x}{data.frame or matrix with data. The last column is assumed to be the total column.}

\item{...}{further arguments passed to \code{q11_1}}

\item{col_fw}{corresponding column to add values with in case the computed sum is less than given by \code{x}. Default to \code{1}.}

\item{cols}{columns to be used to check the sum. It overrides \code{sc} and \code{ec}}
}
\examples{
cpos <- getPosIndex(cnames = names(df1),
                    sc = "P15_1_1_1",
                    ec = "P15_1_5_2") # this (ec) is the column of "total"
x[, cpos] <- q15_1(x = df1, cPos = cpos,
                   nr = 5, nc = 2)
}
