% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveBlanks.R
\name{RemoveBlanks}
\alias{RemoveBlanks}
\title{cut Remove or delete Leading / Trailing Whitespace blanks of a matrix or data.frame with character columns}
\usage{
RemoveBlanks(x)
}
\arguments{
\item{x}{data.frame or matrix}
}
