% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmNonNumeric.R
\name{rmNonNumeric}
\alias{rmNonNumeric}
\title{remove non-numeric symbols. Commas, semicolons, white space between numbers.}
\usage{
rmNonNumeric(x, other = list("oO", ","), fw = list(0, ""))
}
\arguments{
\item{x}{a string vector}

\item{other}{list of extra symbols to be replaced with \code{fw}}

\item{fw}{list. Must be compatible with \code{other}}
}
\value{
a numeric vector or a error mesage if still non-numeric symbols remaining
}
\details{
remove non-numeric symbols as commas, semicolons, white space between numbers, and changes the letters "o" and "O" by zeros (0). String characters coerce them to NA's
}
\examples{
xe <- c("2,000", "4 5    6", "300OOoo", "4.4", "abc", "ab1")
rmNonNumeric(xe)
rmNonNumeric(xe, other = list("oO", ","), fw = list(0, ""))
}
