% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setEmptyCellsValue.R
\name{setEmptyCellsValue}
\alias{setEmptyCellsValue}
\title{Changes multiple invalid values for a unique invalid value}
\usage{
setEmptyCellsValue(x, ef = c(NA, -999, 0), fw = NULL)
}
\arguments{
\item{x}{Vector with multiple empty symbols.}

\item{ef}{Empty Flag number,character, symbol (NA, Inf), etc. What is meant to be "empty". \code{flag} could be a vector but all of itS elements must be different of the posible valid values of \code{x}.}

\item{fw}{The unique number,character, symbol (NA, Inf), etc. which represents empty cells. What is the symbol meaning empty cell?. Default to \code{NA}}
}
\examples{
xe <- c(-999, 1, 0, NA)
print(xe)
setEmptyCellsValue(xe)
setEmptyCellsValue(xe, fw = 1000)
# example 2
xe <- data.frame(a = 3:6, b = 0, c = 0, d = 0, e = 0)
xe[2, 1] <- 0
xe[1, 4] <- 3
xe[2, 3] <- 3
xe[3, 3] <- 3
xe[3, 5] <- 3
xe[4, 2] <- 3
print(xe)
setEmptyCellsValue(x = xe, ef = 0)
}
