% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmode.R
\name{statmode}
\alias{statmode}
\title{computes the mode of data}
\usage{
statmode(x)
}
\arguments{
\item{x}{cathegorical data.}
}
\value{
mode of x.
}
\details{
from http://stats.stackexchange.com/questions/157661/how-to-calculate-mean-median-mode-std-dev-from-distribution
}
\examples{
xe <- rpois(30,12.3)
statmode(x = xe)
names(statmode(x = xe))
}
