% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TotalCompare.R
\name{TotalCompare}
\alias{TotalCompare}
\title{comparison between two vectors}
\usage{
TotalCompare(x, y)
}
\arguments{
\item{x, y}{the two vectors to be compared entry-by-entry.}
}
\value{
a list with
\describe{
  \item{g}{which \code{x} values are greater than \code{y}}
  \item{l}{which \code{x} values are lower   than \code{y}}
}
}
\examples{
xe <- c(1, 3, 6)
ye <- c(1, 2, 7)
TotalCompare(xe, ye)
}
