% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsColsNames.R
\name{xlsColsNames}
\alias{xlsColsNames}
\title{Generates a vector with the characters of excel spreadsheets (A, B, ..., Z, AA, AB,...).}
\usage{
xlsColsNames(n, all = FALSE)
}
\arguments{
\item{n}{The number of columns of the spreadsheet.}

\item{all}{Logical. If \code{length(n) == 1}.}
}
\value{
a character vector of length \code{n} with the excel column index
}
\details{
Generates a vector with the (english) characters of excel spreadsheets (A, B, ..., Z, AA, AB,...).
}
\examples{
xlsColsNames(n = 4)
xlsColsNames(n = c(2,5))
xlsColsNames(n = 60)
}
